Fix to apibaseUrl and database
