+++ src/front/js/constants.js
@@ -1,0 +1,1 @@
+export const apiBaseURL = process.env.BACKEND_URL;
