+++ src/front/js/store/flux.js
@@ -1,6 +1,5 @@
 import { useHistory } from "react-router-dom";
 
-const baseUrl = "https://3001-pink-rodent-ir3cdub5.ws-eu03.gitpod.io/api/";
 const getState = ({ getStore, getActions, setStore }) => {
 	return {
 		store: {
@@ -24,27 +23,7 @@ const getState = ({ getStore, getActions, setStore }) => {
 				getActions().changeColor(0, "green");
 			},
 
-			signUp: user => {
-				let myHeaders = new Headers();
-				myHeaders.append("Content-Type", "application/json");
-				let url = `${baseUrl}sign_up`;
-				let raw = JSON.stringify(user);
-
-				let requestOptions = {
-					method: "POST",
-					headers: myHeaders,
-					body: raw,
-					redirect: "follow"
-				};
-
-				fetch(url, requestOptions)
-					.then(response => response.json())
-					.then(result => {
-						console.log(result);
-					})
-
-					.catch(error => console.log("error", error));
-			},
+			signUp: user => {},
 
 			getMessage: () => {
 				// fetching data from the backend
